feat: WeatherForecastRoute 타입 구현
* WeatherForecastRoute에서 직접 구현하였기에 Period 열거형 제거
* URL과 URLRequest를 구현하는 타입을 별도로 분리하여 재사용성을 높이기 위해 
   Route의 buildRequest(with:) 메서드를 createParameters(latitude:longitude:)로 변경
